crypto_api_test/requirements.txt
behave
requests~=2.32.5
python-decouple